Add sorted function to priority queue and move to utils
