+++ ironfish/src/memPool/index.ts
@@ -3,4 +3,4 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 export * from './memPool'
 export * from './feeEstimator'
-export * from './priorityQueue'
+export * from '../utils/priorityQueue/priorityQueue'
